+++ packages/be-config/src/utils/__tests__/aclExpressionEvaluator.test.ts
@@ -21,7 +21,7 @@ import validMicrolcConfig from '../../__tests__/configurationMocks/validMicrolcC
 describe('Plugins filter tests', () => {
   it('Return all the plugins without expression', () => {
     const plugins = validMicrolcConfig.plugins as Plugin[]
-    const pluginsFiltered = aclExpressionEvaluator(plugins, [])
+    const pluginsFiltered = aclExpressionEvaluator(plugins, [], [])
     expect(pluginsFiltered.length).toBe(plugins.length)
     expect(pluginsFiltered).toMatchObject(plugins)
   })
@@ -39,7 +39,7 @@ describe('Plugins filter tests', () => {
       label: 'Plugin 2',
       aclExpression: '!groups.developer',
     }]
-    const pluginsFiltered = aclExpressionEvaluator(plugins, ['ceo', 'admin', 'developer'])
+    const pluginsFiltered = aclExpressionEvaluator(plugins, ['ceo', 'admin', 'developer'], [])
     expect(pluginsFiltered.length).toBe(1)
     expect(pluginsFiltered[0]).toMatchObject(allowedPlugin)
   })
@@ -61,22 +61,22 @@ describe('Plugins filter tests', () => {
       label: 'Plugin 1',
       aclExpression: 'groups.admin && groups.ceo',
     }]
-    const pluginsFiltered = aclExpressionEvaluator(plugins, ['po', 'reviewer'])
+    const pluginsFiltered = aclExpressionEvaluator(plugins, ['po', 'reviewer'], [])
     expect(pluginsFiltered.length).toBe(2)
     expect(pluginsFiltered).toMatchObject(allowedPlugins)
   })
 
   it('Everything is fine with empty plugins list', () => {
-    const pluginsFiltered = aclExpressionEvaluator([], ['po', 'reviewer'])
+    const pluginsFiltered = aclExpressionEvaluator([], ['po', 'reviewer'], [])
     expect(pluginsFiltered).toMatchObject([])
   })
 
-  it('Everything is fine with empty plugins and groups list', () => {
-    const pluginsFiltered = aclExpressionEvaluator([], [])
+  it('Everything is fine with empty plugins and groups and permissions list', () => {
+    const pluginsFiltered = aclExpressionEvaluator([], [], [])
     expect(pluginsFiltered).toMatchObject([])
   })
 
-  it('general object and no groups', () => {
+  it('general object and no groups and permissions', () => {
     const allConfiguration = {
       test: {
         nested: {
@@ -84,11 +84,11 @@ describe('Plugins filter tests', () => {
         },
       },
     }
-    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [])
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [], [])
     expect(pluginsFiltered).toMatchObject(allConfiguration)
   })
 
-  it('general acl object and no groups', () => {
+  it('general acl object and no groups and permissions', () => {
     const allConfiguration = {
       test: {
         nested: {
@@ -97,11 +97,11 @@ describe('Plugins filter tests', () => {
         },
       },
     }
-    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [])
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [], [])
     expect(pluginsFiltered).toMatchObject({test: {}})
   })
 
-  it('general acl object and groups', () => {
+  it('general acl object and groups but no permissions', () => {
     const allConfiguration = {
       test: {
         nested: {
@@ -110,11 +110,11 @@ describe('Plugins filter tests', () => {
         },
       },
     }
-    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, ['ceo', 'admin'])
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, ['ceo', 'admin'], [])
     expect(pluginsFiltered).toMatchObject(allConfiguration)
   })
 
-  it('general acl object and invalid groups', () => {
+  it('general acl object and invalid groups and no permissions', () => {
     const allConfiguration = {
       test: {
         nested: {
@@ -123,12 +123,38 @@ describe('Plugins filter tests', () => {
         },
       },
     }
-    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, ['po', 'admin'])
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, ['po', 'admin'], [])
+    expect(pluginsFiltered).toMatchObject({test: {}})
+  })
+
+  it('general acl object and invalid permissions', () => {
+    const allConfiguration = {
+      test: {
+        nested: {
+          aclExpression: 'permissions.api.users.get',
+          object: {},
+        },
+      },
+    }
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [], ['api.users.patch'])
+    expect(pluginsFiltered).toMatchObject({test: {}})
+  })
+
+  it('general acl object and permission with less nested object', () => {
+    const allConfiguration = {
+      test: {
+        nested: {
+          aclExpression: 'permissions.api && permissions.api.users && permissions.api.users.get',
+          object: {},
+        },
+      },
+    }
+    const pluginsFiltered = aclExpressionEvaluator(allConfiguration, [], ['api'])
     expect(pluginsFiltered).toMatchObject({test: {}})
   })
 
   it('not object value', () => {
-    const pluginsFiltered = aclExpressionEvaluator(true, ['po', 'admin'])
+    const pluginsFiltered = aclExpressionEvaluator(true, ['po', 'admin'], [])
     expect(pluginsFiltered).toBe(true)
   })
 
@@ -146,8 +172,74 @@ describe('Plugins filter tests', () => {
         },
       ],
     }
-    const filtered = aclExpressionEvaluator(toFilter, ['doctor'])
+    const filtered = aclExpressionEvaluator(toFilter, ['doctor'], [])
     expect(filtered.plugins.length).toBe(1)
+    expect(filtered.plugins.length).not.toBe(toFilter.plugins.length)
+    expect(filtered).not.toBe(toFilter)
+  })
+
+  it('works with both groups and permissions', () => {
+    const toFilter = {
+      plugins: [
+        {
+          aclExpression: 'permissions.api.test-crud.all || permissions.api.test-crud.get',
+        },
+        {
+          aclExpression: 'groups.superadmin || groups.admin || groups.secretary',
+        },
+        {
+          aclExpression: 'groups.superadmin || groups.admin || groups.doctor',
+        },
+        {
+          aclExpression: 'groups.superadmin || groups.admin',
+        },
+      ],
+    }
+    const filtered = aclExpressionEvaluator(toFilter, ['doctor'], ['api.users.get', 'api.users.post', 'api.test-crud.all'])
+    expect(filtered.plugins.length).toBe(2)
+    expect(filtered.plugins.length).not.toBe(toFilter.plugins.length)
+    expect(filtered).not.toBe(toFilter)
+  })
+
+  it('works with permissions with different nested levels', () => {
+    const toFilter = {
+      plugins: [
+        {
+          aclExpression: 'permissions.api',
+        },
+        {
+          aclExpression: 'permissions.api.users.count.get',
+        },
+        {
+          aclExpression: 'permissions.api.companies',
+        },
+      ],
+    }
+    const filtered = aclExpressionEvaluator(toFilter, [], ['api.users.post', 'api.users.count.get'])
+    expect(filtered.plugins.length).toBe(1)
+    expect(filtered.plugins.length).not.toBe(toFilter.plugins.length)
+    expect(filtered).not.toBe(toFilter)
+  })
+
+  it('works with aclExpression with parenthesis and other boolean operators', () => {
+    const toFilter = {
+      plugins: [
+        {
+          aclExpression: '(groups.doctor && !permissions.api.users.post) || permissions.api.users.count.get',
+        },
+        {
+          aclExpression: '(groups.doctor === true && permissions.api.users.post === true)',
+        },
+        {
+          aclExpression: '(groups.doctor === false && permissions.api.users.post === true)',
+        },
+        {
+          aclExpression: '(groups.doctor && permissions.api.users.post === false)',
+        },
+      ],
+    }
+    const filtered = aclExpressionEvaluator(toFilter, ['doctor'], ['api.users.post', 'api.users.count.get'])
+    expect(filtered.plugins.length).toBe(2)
     expect(filtered.plugins.length).not.toBe(toFilter.plugins.length)
     expect(filtered).not.toBe(toFilter)
   })